+++ src/components/SystemsView/TagsModal/TagsModalTable.tsx
@@ -13,7 +13,7 @@ import {
   Spinner,
 } from '@patternfly/react-core';
 import { BulkSelect } from '../../BulkSelect';
-import { System } from '../../InventoryViews/hooks/useHostsQuery';
+import { System } from '../../InventoryViews/hostsQueryOptions';
 import { INITIAL_PAGE, PER_PAGE } from '../../../constants';
 import { NO_HEADER } from '../../InventoryViews/constants';
 import NoEntitiesFound from '../../InventoryTable/NoEntitiesFound';
