+++ src/components/SystemsView/columns/allColumnDefinitions.test.tsx
@@ -2,7 +2,7 @@ import '@testing-library/jest-dom';
 import { render, screen } from '@testing-library/react';
 import React from 'react';
 import allColumns, { type Column } from './allColumnDefinitions';
-import { InventoryViewSystem } from '../../InventoryViews/hooks/useInventoryViewsQuery';
+import { InventoryViewSystem } from '../../InventoryViews/inventoryViewsQueryOptions';
 import {
   DEFAULT_NAME_COLUMN_MIN_WIDTH,
   getColumnMinWidthStyle,
