+++ src/components/SystemsView/columns/allColumnDefinitions.ts
@@ -2,7 +2,7 @@ import { ColumnManagementModalColumn } from '@patternfly/react-component-groups'
 import inventoryColumns from './inventory/columnDefinitions';
 import complianceColumns from './compliance/columnDefinitions';
 import patchColumns from './content/columnDefinitions';
-import { System } from '../../InventoryViews/hooks/useHostsQuery';
+import { System } from '../../InventoryViews/hostsQueryOptions';
 import { Resolve } from '../../../types/utility-types';
 import advisorColumns from './advisor/columnDefinitions';
 import vulnerabilityColumns from './vulnerability/columnDefinitions';
