+++ src/components/SystemsView/columns/compliance/columnDefinitions.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import type { Column } from '../allColumnDefinitions';
-import { InventoryViewSystem } from '../../../InventoryViews/hooks/useInventoryViewsQuery';
+import { InventoryViewSystem } from '../../../InventoryViews/inventoryViewsQueryOptions';
 import { ApiHostViewsGetHostViewsOrderByEnum } from '@redhat-cloud-services/host-inventory-client/ApiHostViewsGetHostViews';
 import LastComplianceScan from './cells/LastComplianceScan';
 import Policies from './cells/Policies';
