+++ src/components/SystemsView/columns/inventory/cells/DisplayName.tsx
@@ -5,7 +5,7 @@ import { Flex, FlexItem, Icon, Popover } from '@patternfly/react-core';
 import { BundleIcon } from '@patternfly/react-icons';
 import FontAwesomeImageIcon from '../../../../FontAwesomeImageIcon';
 import CellValue from '../../CellValue';
-import type { System } from '../../../../InventoryViews/hooks/useHostsQuery';
+import type { System } from '../../../../InventoryViews/hostsQueryOptions';
 
 export type DisplayNameValue = Pick<
   System,
