+++ src/components/SystemsView/columns/inventory/cells/Tags.test.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import type { StructuredTag } from '@redhat-cloud-services/host-inventory-client';
 import Tags from './Tags';
 import { SystemActionModalsContext } from '../../../SystemActionModalsContext';
-import type { System } from '../../../../InventoryViews/hooks/useHostsQuery';
+import type { System } from '../../../../InventoryViews/hostsQueryOptions';
 import { NOT_AVAILABLE } from '../../CellValue';
 
 const SYSTEM_ID = 'test-system-id';
