+++ src/components/SystemsView/hooks/useDeleteSystemsMutation.ts
@@ -2,7 +2,7 @@ import { useMutation } from '@tanstack/react-query';
 import { useAddNotification } from '@redhat-cloud-services/frontend-components-notifications/hooks';
 import { deleteSystemsById } from '../../InventoryTable/utils/api';
 import { getDeleteErrorDescription } from '../../InventoryTable/utils/errorUtils';
-import { type System } from '../../InventoryViews/hooks/useHostsQuery';
+import { type System } from '../../InventoryViews/hostsQueryOptions';
 import { useMemo } from 'react';
 import type { OnInvalidate } from '../SystemActionModalsContext';
